Cleaning up Iterator and Trait
